+++ libs/core/src/utils/Wrapper.ts
@@ -1,7 +1,0 @@
-export type Wrapper = new <T>(plain: T) => T;
-
-export const Wrapper = class Wrapper {
-  public constructor(plain: unknown) {
-    Object.assign(this, plain);
-  }
-} as Wrapper;